fixes for scara driver  (buff o/f, positional errors)
Fixes were made for 3 bugs
1) Potential buffer overflow
2) Accumulative steeping error caused by the motors not being given all
the step commands required to reach the target position.
3) Intermittent fault that caused parameter values to be overwritten
with invalid values.

Upon testing this set of fixes I noticed that my rig still has a very
small non-accumulative error with the laser positioning, which is likely
to be unnoticeable with the pen.  I have yet to track down the "fault"
but it may be due to the motors being driven at 1/16th step. Thus it is
simply a result of the position being at an unstable position in-between
1/4 step positions.  With 1/16th stepping some step positions cannot be
accurately obtained and actual position obtained is a result of the
motion leading up to that point.